+++ lib/hanami/extensions/view/context.rb
@@ -36,11 +36,12 @@ module Hanami
           # @api private
           # @since 2.1.0
           def context_superclass(slice)
-            return Hanami::View::Context if Hanami.app.equal?(slice)
+            # A slice that is its own host builds straight on the framework context
+            return Hanami::View::Context if slice.host?
 
             begin
               slice.inflector.constantize(
-                slice.inflector.camelize("#{slice.app.slice_name.name}/views/context")
+                slice.inflector.camelize("#{slice.host.slice_name.name}/views/context")
               )
             rescue NameError => exception
               raise exception unless %i[Views Context].include?(exception.name)